Classe GameStarter injeta as dependências e inicia jogo
