+++ jogo-oito/src/main/java/chat/gpt/application/Main.java
@@ -1,16 +1,10 @@
 package chat.gpt.application;
 
-import chat.gpt.controller.GameService;
-import chat.gpt.model.Game;
-import chat.gpt.view.GameGUI;
+import chat.gpt.builder.GameStart;
 
 public class Main {
 
 	public static void main(String[] args) {
-		Game game = new Game();
-        GameService service = new GameService(game, null); 
-
-        GameGUI gameGUI = new GameGUI(game, service);
-        service.setView(gameGUI);
+		GameStart.startGame();   
 	}
 }
