+++ jogo-oito/src/main/java/chat/gpt/controller/ButtonActionListener.java
@@ -3,7 +3,7 @@ package chat.gpt.controller;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-import chat.gpt.model.ResetGameButton;
+import chat.gpt.view.ResetGameButton;
 
 public interface ButtonActionListener extends ActionListener {
