+++ jogo-oito/src/main/java/chat/gpt/view/ButtonPiece.java
@@ -1,4 +1,4 @@
-package chat.gpt.model;
+package chat.gpt.view;
 
 import static chat.gpt.util.Constants.DEFAULT_FONT;
